[coroutines-api] Implement async coroutines API: move implementation to more suitable package
