+++ teamcity-rest-client-impl/src/main/kotlin/org/jetbrains/teamcity/rest/coroutines/implementation.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.teamcity.rest
+package org.jetbrains.teamcity.rest.coroutines
 
 import com.google.gson.Gson
 import kotlinx.coroutines.Dispatchers
@@ -16,47 +16,7 @@ import okhttp3.OkHttpClient
 import okhttp3.RequestBody.Companion.toRequestBody
 import okhttp3.ResponseBody
 import okhttp3.logging.HttpLoggingInterceptor
-import org.jetbrains.teamcity.rest.coroutines.ArtifactDependency
-import org.jetbrains.teamcity.rest.coroutines.ArtifactRule
-import org.jetbrains.teamcity.rest.coroutines.Branch
-import org.jetbrains.teamcity.rest.coroutines.Build
-import org.jetbrains.teamcity.rest.coroutines.BuildAgent
-import org.jetbrains.teamcity.rest.coroutines.BuildAgentAuthorizedInfo
-import org.jetbrains.teamcity.rest.coroutines.BuildAgentEnabledInfo
-import org.jetbrains.teamcity.rest.coroutines.BuildAgentLocator
-import org.jetbrains.teamcity.rest.coroutines.BuildAgentPool
-import org.jetbrains.teamcity.rest.coroutines.BuildAgentPoolLocator
-import org.jetbrains.teamcity.rest.coroutines.BuildArtifact
-import org.jetbrains.teamcity.rest.coroutines.BuildCanceledInfo
-import org.jetbrains.teamcity.rest.coroutines.BuildCommentInfo
-import org.jetbrains.teamcity.rest.coroutines.BuildConfiguration
-import org.jetbrains.teamcity.rest.coroutines.BuildLocator
-import org.jetbrains.teamcity.rest.coroutines.BuildProblem
-import org.jetbrains.teamcity.rest.coroutines.BuildProblemOccurrence
-import org.jetbrains.teamcity.rest.coroutines.BuildQueue
-import org.jetbrains.teamcity.rest.coroutines.BuildRunningInfo
-import org.jetbrains.teamcity.rest.coroutines.Change
-import org.jetbrains.teamcity.rest.coroutines.ChangeFile
-import org.jetbrains.teamcity.rest.coroutines.FinishBuildTrigger
-import org.jetbrains.teamcity.rest.coroutines.Investigation
-import org.jetbrains.teamcity.rest.coroutines.InvestigationLocator
-import org.jetbrains.teamcity.rest.coroutines.InvestigationScope
-import org.jetbrains.teamcity.rest.coroutines.Mute
-import org.jetbrains.teamcity.rest.coroutines.MuteLocator
-import org.jetbrains.teamcity.rest.coroutines.Parameter
-import org.jetbrains.teamcity.rest.coroutines.PinInfo
-import org.jetbrains.teamcity.rest.coroutines.Project
-import org.jetbrains.teamcity.rest.coroutines.Revision
-import org.jetbrains.teamcity.rest.coroutines.Test
-import org.jetbrains.teamcity.rest.coroutines.TestLocator
-import org.jetbrains.teamcity.rest.coroutines.TestRun
-import org.jetbrains.teamcity.rest.coroutines.TestRunsLocator
-import org.jetbrains.teamcity.rest.coroutines.TriggeredInfo
-import org.jetbrains.teamcity.rest.coroutines.User
-import org.jetbrains.teamcity.rest.coroutines.UserLocator
-import org.jetbrains.teamcity.rest.coroutines.VcsRoot
-import org.jetbrains.teamcity.rest.coroutines.VcsRootInstance
-import org.jetbrains.teamcity.rest.coroutines.VcsRootLocator
+import org.jetbrains.teamcity.rest.*
 import org.slf4j.LoggerFactory
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
